Refactoring the Key Vault Child ID
