+++ azurerm/resource_arm_key_vault_secret_test.go
@@ -10,117 +10,6 @@ import (
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/utils"
 )
 
-func TestAccAzureRMKeyVaultChild_validateName(t *testing.T) {
-	cases := []struct {
-		Input       string
-		ExpectError bool
-	}{
-		{
-			Input:       "",
-			ExpectError: true,
-		},
-		{
-			Input:       "hello",
-			ExpectError: false,
-		},
-		{
-			Input:       "hello-world",
-			ExpectError: false,
-		},
-		{
-			Input:       "hello-world-21",
-			ExpectError: false,
-		},
-		{
-			Input:       "hello_world_21",
-			ExpectError: true,
-		},
-		{
-			Input:       "Hello-World",
-			ExpectError: false,
-		},
-		{
-			Input:       "20202020",
-			ExpectError: false,
-		},
-		{
-			Input:       "ABC123!@£",
-			ExpectError: true,
-		},
-	}
-
-	for _, tc := range cases {
-		_, errors := validateKeyVaultChildName(tc.Input, "")
-
-		hasError := len(errors) > 0
-
-		if tc.ExpectError && !hasError {
-			t.Fatalf("Expected the Key Vault Child Name to trigger a validation error for '%s'", tc.Input)
-		}
-	}
-}
-
-func TestAccAzureRMKeyVaultSecret_parseID(t *testing.T) {
-	cases := []struct {
-		Input       string
-		Expected    SecretID
-		ExpectError bool
-	}{
-		{
-			Input:       "",
-			ExpectError: true,
-		},
-		{
-			Input:       "https://my-keyvault.vault.azure.net/secrets",
-			ExpectError: true,
-		},
-		{
-			Input:       "https://my-keyvault.vault.azure.net/secrets/bird",
-			ExpectError: true,
-		},
-		{
-			Input:       "https://my-keyvault.vault.azure.net/secrets/bird/fdf067c93bbb4b22bff4d8b7a9a56217",
-			ExpectError: false,
-			Expected: SecretID{
-				Name:            "bird",
-				KeyVaultBaseUrl: "https://my-keyvault.vault.azure.net/",
-				Version:         "fdf067c93bbb4b22bff4d8b7a9a56217",
-			},
-		},
-		{
-			Input:       "https://my-keyvault.vault.azure.net/secrets/bird/fdf067c93bbb4b22bff4d8b7a9a56217/XXX",
-			ExpectError: true,
-		},
-	}
-
-	for _, tc := range cases {
-		secretId, err := parseKeyVaultSecretID(tc.Input)
-		if err != nil {
-			if !tc.ExpectError {
-				t.Fatalf("Got error for ID '%s': %+v", tc.Input, err)
-			}
-
-			return
-		}
-
-		if secretId == nil {
-			t.Fatalf("Expected a SecretID to be parsed for ID '%s', got nil.", tc.Input)
-		}
-
-		if tc.Expected.KeyVaultBaseUrl != secretId.KeyVaultBaseUrl {
-			t.Fatalf("Expected 'KeyVaultBaseUrl' to be '%s', got '%s' for ID '%s'", tc.Expected.KeyVaultBaseUrl, secretId.KeyVaultBaseUrl, tc.Input)
-		}
-
-		if tc.Expected.Name != secretId.Name {
-			t.Fatalf("Expected 'Version' to be '%s', got '%s' for ID '%s'", tc.Expected.Name, secretId.Name, tc.Input)
-		}
-
-		if tc.Expected.Version != secretId.Version {
-			t.Fatalf("Expected 'Version' to be '%s', got '%s' for ID '%s'", tc.Expected.Version, secretId.Version, tc.Input)
-		}
-	}
-}
-
 func TestAccAzureRMKeyVaultSecret_basic(t *testing.T) {
 	resourceName := "azurerm_key_vault_secret.test"
 	rs := acctest.RandString(6)
